Revert "Move createGraphQLClient into library #1810" (#1868)
Revert this as we need a different solution for graphql requests in next
app router. This was never released so it can be removed without issues

New client gets added here:
https://github.com/vivid-planet/comet/pull/1870
