--- .changeset/cuddly-roses-lick.md
@@ -1,8 +1,0 @@
----
-"@comet/cms-site": major
----
-
-Add `createGraphQLClient`
-
-It can be used as a drop-in replacement for the client previously defined in the application code.
-Adds [graphql-request](https://www.npmjs.com/package/graphql-request) as a peer dependency.
