+++ demo/site/package.json
@@ -35,7 +35,7 @@
         "express": "^4.0.0",
         "fs-extra": "^9.0.0",
         "graphql": "^15.0.0",
-        "graphql-request": "^6.1.0",
+        "graphql-request": "^3.0.0",
         "graphql-tag": "^2.12.6",
         "next": "^12.0.0",
         "pure-react-carousel": "^1.0.0",
