+++ demo/site/src/pages/[[...path]].tsx
@@ -1,4 +1,3 @@
-import { PreviewData } from "@comet/cms-site";
 import { defaultLanguage, domain } from "@src/config";
 import { documentTypes } from "@src/documentTypes";
 import NotFound404 from "@src/pages/404";
@@ -9,6 +8,7 @@ import { ParsedUrlQuery } from "querystring";
 import * as React from "react";
 
 import { GQLDocumentTypeQuery, GQLDocumentTypeQueryVariables } from "./[[...path]].generated";
+import { PreviewData } from "./api/site-preview";
 
 type PageProps = {
     documentType: string;
