+++ demo/site/src/util/createGraphQLClient.ts
@@ -1,5 +1,38 @@
-import { createGraphQLClient as libraryCreateGraphQLClient, PreviewData } from "@comet/cms-site";
+import { PreviewData } from "@src/pages/api/site-preview";
+import { GraphQLClient } from "graphql-request";
 
-export default function createGraphQLClient(previewData?: PreviewData) {
-    return libraryCreateGraphQLClient(`${process.env.API_URL_INTERNAL}/graphql`, previewData);
+export default function createGraphQLClient(previewData?: PreviewData): GraphQLClient {
+    const { includeInvisibleBlocks, includeInvisiblePages, previewDamUrls } = {
+        includeInvisiblePages: !!previewData,
+        includeInvisibleBlocks: previewData && previewData.includeInvisible,
+        previewDamUrls: !!previewData,
+    };
+
+    const headers: Record<string, string> = {};
+
+    const includeInvisibleContentHeaderEntries: string[] = [];
+
+    if (includeInvisiblePages) {
+        includeInvisibleContentHeaderEntries.push("Pages:Unpublished");
+    }
+
+    if (includeInvisibleBlocks) {
+        includeInvisibleContentHeaderEntries.push("Blocks:Invisible");
+    }
+
+    // tells api to send invisble content
+    // authentication is required when this header is used
+    if (includeInvisibleContentHeaderEntries.length > 0) {
+        headers["x-include-invisible-content"] = includeInvisibleContentHeaderEntries.join(",");
+    }
+
+    // tells api to create preview image urls
+    // authentication is required when this header is used
+    if (previewDamUrls) {
+        headers["x-preview-dam-urls"] = "1";
+    }
+
+    return new GraphQLClient(`${process.env.API_URL_INTERNAL}/graphql`, {
+        headers,
+    });
 }
