+++ packages/site/cms-site/package.json
@@ -41,8 +41,6 @@
         "@types/styled-components": "^5.0.0",
         "chokidar-cli": "^2.0.0",
         "eslint": "^8.0.0",
-        "graphql": "^15.0.0",
-        "graphql-request": "^6.0.0",
         "jest": "^29.5.0",
         "jest-environment-jsdom": "^29.5.0",
         "jest-junit": "^15.0.0",
@@ -57,7 +55,6 @@
         "typescript": "^4.0.0"
     },
     "peerDependencies": {
-        "graphql-request": "3 - 6",
         "next": "^12.0.0",
         "react": "^17.0.0",
         "react-dom": "^17.0.0",
